Setup the Stage and Player Movement
DONE- Move the bounce pointer to the beginning of the course

DONE- match the start and finish flags to the rest of the course

DONE-Put the course in a folder

DONE-Got rid of these "New Game Objects" that kept poping up

DONE-Set the position of the player, as well as putting it over the ground sprite using a Sorting Layer

DONE-Registered Input to work

A = 1
B = 2
X = 3
Y = 4
L = 5
R = 6
FreeSpace = 7
Start = 8
Finish = 9

if a player presses the correct button in the next slot, move the player forward.

>Keep track of what spot is in the next slot
>Keep track of what it's "button" is, which is an integer
>When a player presses a button, have that button register as a certain integer
>Then, compare that integer the player pressed to the integer in the next slot.
>If they are the same, move the bouncepointer to that slot, and increment the "next slot" to keep track of

---When you come Back
DONE >Figure out how to register multiple inputs in a single leap. As well as how to turn back on the ability to read inputs.

DONE>Have the player continue to move from square to square after every correct input. For some reason, it registers a correct input as an incorrect input.

DONE> Have the Camera follow the player, and make sure to be able to skip over empty spots

DONE> Have the player jump over gaps by pressing the next correct button

DONE> Decreased the size of the floor hitbox so the player doesn't bounce off of it
DONE>Player bounces from one spot to another
DONE> Player bounces while storing paths during input. Automatically jumping again after a jump path has finished.