Camera Follow, Tripping, Player Progress Bar, Multiplayer Scene - Implemented
DONE> Have the camera adjust from one side of the stage to another
(set XMin to 0, and XMax to Z, where z is some fuction of the length of the path array... no actually, just have it be establshed via stage information)

Example:
xmin = 0
xmax = 68.15

DONE> Tripping Implemented. If you press the wrong button next, then your inputs are stunned until you land and wait a bit

DONE> Given consecutive successful bounces, let there be an int that increases. The higher the int, the higher the speed, up to a max speed. When an incorrect input is given, reset the counter and speed to regular levels.

DONE> Finished ProgressBar

DONE> Multiplayer Scene. Two players can play in the Multiplayer scene