Animation Setup Complete
DONE: Refined the Camera and Scaling in HopnScotch. Leaving the Animator Tool open slows down the game (Besides the Canvas). Don't mind it.
DONE: Complete the Second Color's Animations so that you can create a database to store them in and try to have the game grab them and put them onto players
DONE: Players can access their respective animations depending on the name of the animationcontroller to retrieve in the AnimationEncyclopedia
This works in both Single and Multiplayer