Spot Prefabs and Path Graphic Layout DONE
FIXED and DONE
Created a bunch of prefabs with the different sprites and object combinations
Adjusted the size of the sprites of the floor sprites so that they are parallel
The Path is an array of CHARs, not INTs. Made the adjustment to the code.
Adjusted the height of the start and finish to
The Sprites changes dynamically as the path is being formed.