Menus Work. Button Config works. Clouds drift left and right. GameManager exists only in Game Scenes. Game... works
There being two Player Inputs in the scene seems to be the problem.

Okay so, now the menus work. The players can cycle through all the scenes well. The Buttons can register too
The GameManager spawns only in the GameScene and is deleted when exiting the GameScene (Resetting)
Button Configuration Settings persist through the scenes