Hop n' Scotch Revamped
Traditional Movement Implemented
Jumping Implemented
Wall Jumping Implemented
Go through platforms implemented

-------------------------------------------------------

Part 1:
Lock On works so far
Pressing L turns on the LockOnZone
The LockOnZone detects GameObjects that are in the Character and Targets Layer
The LockOnZone follows the player
The LockOnZone size can be adjusted in the UnityEditor, Both the Gizmo, the Graphic, and the actual Physics2D Area

-----------------
Part 2:
I made a collider to register whether or not it sees a target. If it does, the target, onCollissionEnter, will check to see if the collider's tag is "Player". if so, the target will reveal it's appropriate retacle.

Make sure to attach the correct rectacles to the correct labels, and thus to the correct targets

The Rectangle
Collider Trigger (For triggering the rectacle in each target to appear
For the Graphical Representation of the lock on area
for the LockOnTech Gizmo to track what colliders are in the area
Now changes in real time. This way, the lock on doesn't blink into existance. Apparently, blinking into existance without moving doesn't trigger the OnTriggerEnter function, so this is a nifty workaround

I made several colliders and results arrays, one for each button, A,B,X,Y, and
gave each of them their own Physics2D.OverlapAreaNonAlloc with an array of size 2 for each button. This way, we can have four different checks tracking four different types of targets.

Also, each target, after leaving the trigger collision, calls a function to the player's LockOnTech Script, running the EscapeLock function, thereby removing itself from the lost of lock on targets the player can home in on.

Just make sure that a target is in the correct layer for the LockOnTech's four  Physics2D.OverlapAreaNonAlloc scanners to track that target appropriately

Also, make sure the targets have only ONE Collider total. The Target's colliders do not have to be triggers. The LockOn is a trigger that locks on to their non-trigger colliders. Now you're not getting Debug.Log calls twice in a row, also targets can slip in and out of the target zone seemlessly with being able to leave their respective target_results array in the LockOn Tech script of the Player.
-----------------

Part 3:
Coded All night on HnS
Homing Attack Almost complete
Distinguishes between four labels A,B,X, and Y
Prioritizes the closest target
You slow down in the air in LockOn Mode
You can LockOn in midair only once, unless you hit your target, in which case, you can chain Homing Hops
If you hit the wall or Floor, your Homing Hop Cancels

Give Targets Trigger Colliders, it works fine. Make their colliders if they need to stand, super low.
------------------------------

Part 4:

Colliders Set Up, Base,Dash,Crouch,Homing, can switch fluently
Can Crawl correctly
Damage Calculation for Targets is Done. With that, Homing Hop System Mechanic is completely done (as far as I know) Will need to test what happens if you get intercepted by a different target